lowercased the first char of a lot of errors, catch create mute role errors

+++ SlashCommands/Moderation/mute.js
@@ -50,7 +50,7 @@ module.exports = {
                 type: 'mute'
             });
 
-            if (hasMuteRecord) return client.util.throwError(interaction, 'This user already currently muted');
+            if (hasMuteRecord) return client.util.throwError(interaction, 'this user already currently muted');
 
             new Infraction(client, 'Mute', interaction, interaction.member, user, {
                 reason: reason,
@@ -98,12 +98,14 @@ module.exports = {
         });
 
         const role = interaction.guild.roles.cache.get(muterole) || (await client.util.createMuteRole(interaction));
+        if (!role) return client.util.throwError(interaction, 'the muted role was not found, so I tried to create one, but I failed due to bad permissions');
+        if (role.id !== muterole) client.cache.settings.delete(interaction.guild.id);
 
         if (role.position >= interaction.guild.me.roles.highest.position)
-            return client.util.throwError(interaction, 'My hierarchy is too low to manage the muted role');
+            return client.util.throwError(interaction, 'my hierarchy is too low to manage the muted role');
 
         if (member.roles.cache.has(role.id))
-            return client.util.throwError(interaction, 'This user is already currently muted!');
+            return client.util.throwError(interaction, 'this user is already currently muted!');
         else if (hasMuteRecord) {
             await punishmentSchema.deleteMany({
                 guildID: interaction.guild.id,
